jules prompt/requirements.txt
python-telegram-bot>=20.0
pytest
pytest-asyncio
aiosqlite
